+++ backend/src/main/java/io/github/changmin6362/multistore/feature/authorization/controller/RoleController.java
@@ -1,17 +1,17 @@
 package io.github.changmin6362.multistore.feature.authorization.controller;
 
 import io.github.changmin6362.multistore.common.web.ApiResponse;
-import io.github.changmin6362.multistore.feature.authorization.web.response.RoleResponse;
+import io.github.changmin6362.multistore.common.web.flags.DeletedResponse;
 import io.github.changmin6362.multistore.feature.authorization.service.RoleService;
+import io.github.changmin6362.multistore.feature.authorization.web.request.RoleCreateRequest;
+import io.github.changmin6362.multistore.feature.authorization.web.request.RoleUpdateRequest;
+import io.github.changmin6362.multistore.feature.authorization.web.response.RoleResponse;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.util.StringUtils;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
-import io.github.changmin6362.multistore.feature.authorization.web.request.RoleCreateRequest;
-import io.github.changmin6362.multistore.feature.authorization.web.request.RoleUpdateRequest;
-import io.github.changmin6362.multistore.common.web.flags.DeletedResponse;
 
 @RestController
 @RequestMapping("/api/rbac/roles")
@@ -69,7 +69,7 @@ public class RoleController {
      * 역할 단건 조회 (없으면 404)
      */
     @GetMapping("/{roleId}")
-    public ResponseEntity<ApiResponse> getRole(@PathVariable Long roleId) {
+    public ResponseEntity<ApiResponse> getRole(@PathVariable int roleId) {
         RoleResponse role = roleService.findById(roleId);
         if (role == null) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND)
@@ -85,8 +85,8 @@ public class RoleController {
      * 요청 바디: { "roleName": "ROLE_MANAGER", "roleDescription": "..." }
      */
     @PutMapping("/{roleId}")
-    public ResponseEntity<ApiResponse> updateRole(@PathVariable Long roleId,
-                                                                       @RequestBody RoleUpdateRequest body) {
+    public ResponseEntity<ApiResponse> updateRole(@PathVariable int roleId,
+                                                  @RequestBody RoleUpdateRequest body) {
         String roleName = body != null ? body.roleName() : null;
         String roleDescription = body != null ? body.roleDescription() : null;
         if (!StringUtils.hasText(roleName)) {
@@ -108,7 +108,7 @@ public class RoleController {
      * 역할 삭제 (없으면 404)
      */
     @DeleteMapping("/{roleId}")
-    public ResponseEntity<ApiResponse> deleteRole(@PathVariable Long roleId) {
+    public ResponseEntity<ApiResponse> deleteRole(@PathVariable int roleId) {
         boolean deleted = roleService.delete(roleId);
         if (!deleted) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND)
