+++ backend/src/main/java/io/github/changmin6362/multistore/feature/authorization/web/response/PermissionResponse.java
@@ -10,7 +10,7 @@ package io.github.changmin6362.multistore.feature.authorization.web.response;
  * @param actionType
  */
 public record PermissionResponse(
-        Long permissionId,
+        int permissionId,
         String permissionName,
         String permissionDescription,
         String resourceType,
